Revert "Vertical layout"
