+++ src/Configuration.ts
@@ -21,11 +21,6 @@ export enum AutoExpand {
   None,
 }
 
-export enum Layout {
-  Horizontal,
-  Vertical,
-}
-
 export interface Configuration {
   providerFactory: ProviderFactory;
   providerParams: { [key: string]: string };
@@ -36,7 +31,6 @@ export interface Configuration {
   refreshInterval: number;
   showHourlyPrecipitation: boolean;
   showHourlyWind: boolean;
-  layout: Layout;
 }
 
 const DEFAULT_CONFIGURATION: Configuration = {
@@ -49,7 +43,6 @@ const DEFAULT_CONFIGURATION: Configuration = {
   refreshInterval: 2 * 3600,
   showHourlyPrecipitation: true,
   showHourlyWind: true,
-  layout: Layout.Horizontal,
 };
 
 function decodeConfiguration(params: { [key: string]: string }): Configuration {
@@ -70,7 +63,6 @@ function decodeConfiguration(params: { [key: string]: string }): Configuration {
   const showHourlyPrecipitation =
     params['hourly_precipitation'] === undefined ? DEFAULT_CONFIGURATION.showHourlyPrecipitation : params['hourly_precipitation'] === 'true' ? true : false;
   const showHourlyWind = params['hourly_wind'] === undefined ? DEFAULT_CONFIGURATION.showHourlyWind : params['hourly_wind'] === 'true' ? true : false;
-  const layout = params['layout'] === 'vertical' ? Layout.Vertical : Layout.Horizontal;
 
   return {
     providerFactory,
@@ -82,7 +74,6 @@ function decodeConfiguration(params: { [key: string]: string }): Configuration {
     refreshInterval,
     showHourlyPrecipitation,
     showHourlyWind,
-    layout,
   };
 }
 
@@ -116,9 +107,7 @@ function encodeConfiguration(configuration: Configuration): object {
   if (configuration.showHourlyWind !== DEFAULT_CONFIGURATION.showHourlyWind) {
     params['hourly_wind'] = configuration.showHourlyWind.toString();
   }
-  if (configuration.layout !== DEFAULT_CONFIGURATION.layout) {
-    params['layout'] = configuration.layout === Layout.Vertical ? 'vertical' : 'horizontal';
-  }
+
   return params;
 }
