+++ src/components/CurrentDetails.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { CurrentWeather, SunTimes, DailyPrecipitation } from '../providers/Provider';
+  import type { CurrentWeather } from '../providers/Provider';
 
   import Conditions from './scalars/Conditions.svelte';
   import ConditionsIcon from './scalars/ConditionsIcon.svelte';
@@ -9,10 +9,8 @@
   import UVIndex from './scalars/UVIndex.svelte';
   import Distance from './scalars/Distance.svelte';
   import Pressure from './scalars/Pressure.svelte';
-  import DailySunDetails from './DailySunDetails.svelte';
 
   export let current: CurrentWeather;
-  export let suntimes: SunTimes;
 </script>
 
 <div>
@@ -38,7 +36,5 @@
       <div><span class="font-semibold">Visibility: </span><Distance value={current.visibility} /></div>
     {/if}
     <div><span class="font-semibold">Pressure: </span><Pressure value={current.pressure} /></div>
-    <DailySunDetails {suntimes} />
-    <!-- <DailyPrecipitationDetails {precipitation} /> -->
   </div>
 </div>
