+++ src/components/HourlyDetails.svelte
@@ -1,17 +1,13 @@
 <script lang="ts">
   import { ConditionsIcon } from '../providers/Provider';
   import type { HourlyWeather } from '../providers/Provider';
-  import { configuration, Layout } from '../Configuration';
 
-  import HourlyVerticalLayout from './HourlyVerticalLayout.svelte';
-  import HourlyHorizontalLayout from './HourlyHorizontalLayout.svelte';
   import Temperature from './scalars/Temperature.svelte';
   import Timestamp from './scalars/Timestamp.svelte';
 
   /* Properties */
 
   export let hourly: HourlyWeather[] = [];
-  export let today: boolean = false;
 
   /* Constants */
 
@@ -66,6 +62,8 @@
     [ConditionsIcon.Thunderstorm]: HourlyConditions.Rain,
   };
 
+  const MIN_TEMPERATURE_OPACITY = 0.25;
+
   /* State */
 
   let aggregation: {
@@ -74,7 +72,12 @@
   }[] = [];
   let temperatureLow: number;
   let temperatureHigh: number;
-  let selectedPill = 'Temp';
+
+  function temperatureOpacity(value: number): string {
+    let percent = (value - temperatureLow) / (temperatureHigh - temperatureLow);
+    const opacity = percent * (1 - MIN_TEMPERATURE_OPACITY) + MIN_TEMPERATURE_OPACITY;
+    return opacity.toFixed(2);
+  }
 
   $: {
     /* Aggregate conditions into contiguous regions */
@@ -94,8 +97,53 @@
   }
 </script>
 
-{#if $configuration.layout === Layout.Horizontal}
-  <HourlyHorizontalLayout {aggregation} {hourly} {temperatureLow} {temperatureHigh} {CLASS_TEXT_MAP} />
-{:else if $configuration.layout === Layout.Vertical}
-  <HourlyVerticalLayout {aggregation} {today} {hourly} {temperatureLow} {temperatureHigh} {CLASS_TEXT_MAP} />
-{/if}
+<div class="-mb-2">
+  <div class="flex mb-1.5 overflow-hidden rounded-md text-sm sm:text-base">
+    {#each aggregation as entry}
+      <div class="h-10 leading-9 {CLASS_TEXT_MAP[entry.conditions][0]} text-center" style="width: {(100 * entry.duration) / 24}%;">
+        {#if entry.duration > 4}
+          <div class="truncate">{CLASS_TEXT_MAP[entry.conditions][1]}</div>
+        {:else if entry.duration > 2}
+          <div class="hidden md:block truncate">{CLASS_TEXT_MAP[entry.conditions][1]}</div>
+        {/if}
+      </div>
+    {/each}
+  </div>
+  <div class="flex w-full mb-1">
+    {#each Array(25) as _, i}
+      <div class="{i % 2 === 0 ? 'h-[8px]' : 'h-[5px]'} border-l border-gray-400" style="width: {i < 24 ? 100 / 24 : 0}%;" />
+    {/each}
+  </div>
+  <div class="flex w-full text-xs sm:text-sm">
+    <div style="width: {100 / 24}%;">
+      <div class="hidden md:block">
+        <Timestamp value={hourly[0].timestamp} format="hour" />
+      </div>
+    </div>
+    {#each Array(11) as _, i}
+      {@const timestamp = hourly[2 * (i + 1)].timestamp}
+      <div class="text-center" style="width: {100 / 12}%;">
+        <div class="{i === 0 || i === 2 || i === 4 || i === 6 || i === 8 || i === 10 ? 'block' : 'hidden'} md:block">
+          <Timestamp value={timestamp} format="hour" />
+        </div>
+      </div>
+    {/each}
+    <div style="width: {100 / 24}%;" />
+  </div>
+  <div class="flex w-full text-base sm:text-lg font-light text-black dark:text-white">
+    <div style="width: {100 / 24}%; opacity: {temperatureOpacity(hourly[0].temperature)};">
+      <div class="hidden md:block">
+        <Temperature value={hourly[0].temperature} />
+      </div>
+    </div>
+    {#each Array(11) as _, i}
+      {@const temperature = hourly[2 * (i + 1)].temperature}
+      <div class="text-center" style="width: {100 / 12}%; opacity: {temperatureOpacity(temperature)};">
+        <div class="{i === 0 || i === 2 || i === 4 || i === 6 || i === 8 || i === 10 ? 'block' : 'hidden'} md:block">
+          &nbsp;<Temperature value={temperature} />
+        </div>
+      </div>
+    {/each}
+    <div style="width: {100 / 24}%;" />
+  </div>
+</div>
