+++ src/providers/Provider.ts
@@ -66,16 +66,6 @@ export interface HourlyWeather {
   precipitation_type?: PrecipitationType;
 }
 
-export interface SunTimes {
-  sunrise_timestamp: Date;
-  sunset_timestamp: Date;
-}
-
-export interface DailyPrecipitation {
-  precipitation_probability?: number; // in integral percent 0-100
-  precipitation_amount?: number; // in mm
-}
-
 export interface Weather {
   current: CurrentWeather;
   daily: DailyWeather[];
